Add tsconfig.json file and update .gitignore
Introduced the tsconfig.json file to transition the project to TypeScript, aiming at a more type-safe and maintainable codebase. The 'target' setting is set as ES2017, 'module' as 'commonjs', and 'esModuleInterop' is enabled for better CommonJS module support. Also, 'moduleResolution' is set to 'node' with the 'baseUrl' configured as './'. Custom path '@/*' mapping to 'src/*' is also added for better import statements. 'outDir' is configured to be './dist'. All comments are removed from the output files. Lastly, 'strict' mode is enabled for strict type-checking and lib checking is skipped.

.gitignore file is also updated to exclude .idea directory - the directory for JetBrains IDEs like WebStorm.