Improved http-websocket protocol mix and added multipart/form-data handling
Messages to the server are sent via http protocol, while data from the server is received via webSocket protocol. This implementation will allow to send to the server including larges binary files, while providing a higher level of security. In turn, the transfer and processing of files via multipart and their uploading to S3 is implemented.