Add fileName field and refactor S3 operations in code base
A 'fileName' field has been added to the 'Message' table in the Prisma database. Also, the S3 related codebase has been significantly refactored. The operations are now encapsulated within a newly created S3 class that includes both upload and delete functionality. On top of this, the upload function further secures the file by generating a unique uuid for each upload.