Remove custom methods and add fileName to sendMessage function
The custom methods 'verifyJWT' and 'verifyUserAndPassword' have been removed from the 'sendMessage' function and the type of fastify has been changed from 'MyFastifyInstance' to 'FastifyInstance'. A new field 'fileName' is included in the function to carry the name of the file being sent. It is also added to the data part of the message.