Replace global websocket variable with connections array
The 'ws' global variable in the webSocket and sendMessage modules has been replaced with an array named 'connections'. This approach allows preserving instances of all active connections which makes it possible to broadcast messages to all connected clients instead of just the latest one. Also, deletion of a message is now broadcasted to all connections.